[REF] Organizando as views e models de 'br_account'
