+++ br_account/__manifest__.py
@@ -21,12 +21,22 @@
         'data/account_title_type.xml',
         'views/account_financial_operation.xml',
         'views/account_title_type.xml',
-        'views/account_fiscal_position_view.xml',
-        'views/account_invoice_view.xml',
-        'views/account.xml',
-        'views/br_account_view.xml',
-        'views/product_view.xml',
-        'views/res_company_view.xml',
+        'views/account_fiscal_position.xml',
+        'views/account_move_line.xml',
+        'views/account_invoice.xml',
+        'views/account_invoice_line.xml',
+        'views/account_payment.xml',
+        'views/account_account.xml',
+        'views/br_account.xml',
+        'views/br_account_cfop.xml',
+        'views/br_account_cnae.xml',
+        'views/br_account_document_serie.xml',
+        'views/br_account_fiscal_document.xml',
+        'views/br_account_fiscal_observation.xml',
+        'views/br_account_import_declaration_line.xml',
+        'views/br_account_service_type.xml',
+        'views/product_template.xml',
+        'views/res_company.xml',
         'views/account_tax.xml',
         'views/product_fiscal_classification.xml',
         'wizard/br_product_fiscal_classification_wizard.xml',
