+++ br_account/models/__init__.py
@@ -2,17 +2,31 @@
 # © 2016 Danimar Ribeiro, Trustcode
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import cst
-from . import account
+from . import account_account
+from . import account_chart_template
 from . import account_financial_operation
-from . import account_title_type
-from . import br_account
-from . import product_fiscal_classification
-from . import product
-from . import res_company
 from . import account_fiscal_position
+from . import account_fiscal_position_tax_rule
 from . import account_invoice
 from . import account_invoice_line
-from . import account_tax
 from . import account_move
+from . import account_move_line
+from . import account_payment_term
+from . import account_tax
+from . import account_tax_template
+from . import account_title_type
+from . import br_account_cfop
+from . import br_account_cnae
+from . import br_account_document_related
+from . import br_account_document_serie
+from . import br_account_fiscal_document
+from . import br_account_fiscal_observation
+from . import br_account_import_declaration
+from . import br_account_import_declaration_line
+from . import br_account_invoice_parcel
+from . import br_account_service_type
+from . import cst
+from . import product_fiscal_classification
+from . import product_template
+from . import res_company
 from . import res_partner
