+++ br_account/models/account_invoice.py
@@ -3,9 +3,10 @@
 # © 2016 Danimar Ribeiro, Trustcode
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from odoo import api, fields, models, _
-from odoo.exceptions import UserError
+from odoo import api, fields, models
+from odoo.tools.translate import _
 from odoo.addons import decimal_precision as dp
+from odoo.exceptions import UserError
 from odoo.tools import float_compare
