+++ br_account/models/account_tax.py
@@ -6,60 +6,6 @@
 from odoo import api, fields, models
 
 
-class AccountChartTemplate(models.Model):
-    _inherit = 'account.chart.template'
-
-    @api.multi
-    def _load_template(self, company, code_digits=None,
-                       transfer_account_id=None, account_ref=None,
-                       taxes_ref=None):
-        acc_ref, tax_ref = super(AccountChartTemplate, self)._load_template(
-            company, code_digits, transfer_account_id, account_ref, taxes_ref)
-
-        tax_tmpl_obj = self.env['account.tax.template']
-        tax_obj = self.env['account.tax']
-        for key, value in tax_ref.items():
-            tax_tmpl_id = tax_tmpl_obj.browse(key)
-            tax_obj.browse(value).write({
-                'deduced_account_id': acc_ref.get(
-                    tax_tmpl_id.deduced_account_id.id, False),
-                'refund_deduced_account_id': acc_ref.get(
-                    tax_tmpl_id.refund_deduced_account_id.id, False)
-            })
-        return acc_ref, tax_ref
-
-
-class AccountTaxTemplate(models.Model):
-    _inherit = 'account.tax.template'
-
-    deduced_account_id = fields.Many2one(
-        'account.account.template', string=u"Conta de Dedução da Venda")
-    refund_deduced_account_id = fields.Many2one(
-        'account.account.template', string=u"Conta de Dedução do Reembolso")
-    domain = fields.Selection([('icms', 'ICMS'),
-                               ('icmsst', 'ICMS ST'),
-                               ('simples', 'Simples Nacional'),
-                               ('pis', 'PIS'),
-                               ('cofins', 'COFINS'),
-                               ('ipi', 'IPI'),
-                               ('issqn', 'ISSQN'),
-                               ('ii', 'II'),
-                               ('icms_inter', u'Difal - Alíquota Inter'),
-                               ('icms_intra', u'Difal - Alíquota Intra'),
-                               ('fcp', 'FCP'),
-                               ('csll', 'CSLL'),
-                               ('irrf', 'IRRF'),
-                               ('inss', 'INSS'),
-                               ('outros', 'Outros')], string="Tipo")
-    amount_type = fields.Selection(selection_add=[('icmsst', 'ICMS ST')])
-
-    def _get_tax_vals(self, company):
-        res = super(AccountTaxTemplate, self)._get_tax_vals(company)
-        res['domain'] = self.domain
-        res['amount_type'] = self.amount_type
-        return res
-
-
 class AccountTax(models.Model):
     _inherit = 'account.tax'
