+++ br_account/models/br_account_cnae.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+# © 2009 Renato Lima - Akretion
+# © 2014  KMEE - www.kmee.com.br
+# © 2016 Danimar Ribeiro, Trustcode
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+
+from odoo import api, fields, models
+
+
+class BrAccountCNAE(models.Model):
+    _name = 'br_account.cnae'
+    _description = 'Cadastro de CNAE'
+
+    code = fields.Char(u'Código', size=16, required=True)
+    name = fields.Char(u'Descrição', size=64, required=True)
+    version = fields.Char(u'Versão', size=16, required=True)
+    parent_id = fields.Many2one('br_account.cnae', 'CNAE Pai')
+    child_ids = fields.One2many(
+        'br_account.cnae', 'parent_id', 'CNAEs Filhos')
+    internal_type = fields.Selection(
+        [('view', u'Visualização'), ('normal', 'Normal')],
+        'Tipo Interno', required=True, default='normal')
+
+    @api.model
+    def name_search(self, name, args=None, operator='ilike', limit=100):
+        args = args or []
+        recs = self.browse()
+        if name:
+            recs = self.search([('code', operator, name)] + args, limit=limit)
+        if not recs:
+            recs = self.search([('name', operator, name)] + args, limit=limit)
+        return recs.name_get()
+
+    @api.multi
+    def name_get(self):
+        result = []
+        for rec in self:
+            result.append((rec.id, "%s - %s" % (rec.code, rec.name or '')))
+        return result
