+++ br_account/models/br_account_fiscal_document.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+# © 2009 Renato Lima - Akretion
+# © 2014  KMEE - www.kmee.com.br
+# © 2016 Danimar Ribeiro, Trustcode
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+
+from odoo import fields, models
+
+
+class BrAccountFiscalDocument(models.Model):
+    _name = 'br_account.fiscal.document'
+    _description = 'Tipo de Documento Fiscal'
+
+    code = fields.Char(u'Codigo', size=8, required=True)
+    name = fields.Char(u'Descrição', size=64)
+    electronic = fields.Boolean(u'Eletrônico')
