+++ br_account/models/br_account_import_declaration.py
@@ -1,0 +1,52 @@
+# -*- coding: utf-8 -*-
+# © 2009 Renato Lima - Akretion
+# © 2014  KMEE - www.kmee.com.br
+# © 2016 Danimar Ribeiro, Trustcode
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+
+from odoo import fields, models
+from odoo.addons import decimal_precision as dp
+
+
+class BrAccountImportDeclaration(models.Model):
+    _name = 'br_account.import.declaration'
+
+    invoice_line_id = fields.Many2one(
+        'account.invoice.line', u'Linha de Documento Fiscal',
+        ondelete='cascade', index=True)
+    name = fields.Char(u'Número da DI', size=10, required=True)
+    date_registration = fields.Date(u'Data de Registro', required=True)
+    state_id = fields.Many2one(
+        'res.country.state', u'Estado',
+        domain="[('country_id.code', '=', 'BR')]", required=True)
+    location = fields.Char(u'Local', required=True, size=60)
+    date_release = fields.Date(u'Data de Liberação', required=True)
+    type_transportation = fields.Selection([
+        ('1', u'1 - Marítima'),
+        ('2', u'2 - Fluvial'),
+        ('3', u'3 - Lacustre'),
+        ('4', u'4 - Aérea'),
+        ('5', u'5 - Postal'),
+        ('6', u'6 - Ferroviária'),
+        ('7', u'7 - Rodoviária'),
+        ('8', u'8 - Conduto / Rede Transmissão'),
+        ('9', u'9 - Meios Próprios'),
+        ('10', u'10 - Entrada / Saída ficta'),
+    ], u'Transporte Internacional', required=True, default="1")
+    afrmm_value = fields.Float(
+        'Valor da AFRMM', digits=dp.get_precision('Account'), default=0.00)
+    type_import = fields.Selection([
+        ('1', u'1 - Importação por conta própria'),
+        ('2', u'2 - Importação por conta e ordem'),
+        ('3', u'3 - Importação por encomenda'),
+    ], u'Tipo de Importação', default='1', required=True)
+    thirdparty_cnpj = fields.Char('CNPJ', size=18)
+    thirdparty_state_id = fields.Many2one(
+        'res.country.state', u'Estado',
+        domain="[('country_id.code', '=', 'BR')]")
+    exporting_code = fields.Char(
+        u'Código do Exportador', required=True, size=60)
+    line_ids = fields.One2many(
+        'br_account.import.declaration.line',
+        'import_declaration_id', 'Linhas da DI')
