+++ br_account/models/product_template.py
@@ -2,9 +2,11 @@
 # © 2009  Gabriel C. Stabel
 # © 2009  Renato Lima - Akretion
 # © 2016 Danimar Ribeiro, Trustcode
+# © 2017 Michell Stuttgart, MultidadosTI
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from odoo import models, fields, api
+from odoo import api, fields, models
+
 from .cst import ORIGEM_PROD
