+++ br_account/models/res_company.py
@@ -3,7 +3,7 @@
 # © 2016 Danimar Ribeiro, Trustcode
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from odoo import models, fields
+from odoo import fields, models
 from odoo.addons import decimal_precision as dp
 
 COMPANY_FISCAL_TYPE = [
