+++ br_account/views/account_account.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="view_account_form" model="ir.ui.view">
+        <field name="name">br_account.account.account.form</field>
+        <field name="model">account.account</field>
+        <field name="inherit_id" ref="account.view_account_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='code']" position="after">
+                <field name="code_first_digit"/>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
