+++ br_account/views/account_payment.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="br_account_payment_term_form" model="ir.ui.view">
+        <field name="name">br_account.payment.term.form</field>
+        <field name="model">account.payment.term</field>
+        <field name="inherit_id" ref="account.view_payment_term_form"/>
+        <field name="arch" type="xml">
+            <field name="active" position="after">
+                <field name="indPag" required="1"/>
+            </field>
+        </field>
+    </record>
+
+    <record id="action_br_account_account_invoice_payment" model="ir.actions.act_window">
+        <field name="name">Register Payment</field>
+        <field name="res_model">account.payment</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">form</field>
+        <field name="view_id" ref="account.view_account_payment_invoice_form"/>
+        <field name="target">new</field>
+    </record>
+
+
+</odoo>
