+++ br_account/views/br_account_fiscal_document.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record id="view_br_account_fiscal_document_form" model="ir.ui.view">
+        <field name="name">br_account.fiscal.document</field>
+        <field name="model">br_account.fiscal.document</field>
+        <field name="arch" type="xml">
+            <form string="Documento Fiscal">
+                <sheet>
+                    <group>
+                        <field name="code" select="1"/>
+                        <field name="name" select="1"/>
+                        <field name="electronic" select="2"/>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record id="view_br_account_fiscal_document_tree" model="ir.ui.view">
+        <field name="name">br_account.fiscal.document</field>
+        <field name="model">br_account.fiscal.document</field>
+        <field name="arch" type="xml">
+            <tree string="Documento Fiscal">
+                <field name="code" select="1"/>
+                <field name="name" select="1"/>
+                <field name="electronic" select="2"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="action_br_account_fiscal_document_form" model="ir.actions.act_window">
+        <field name="name">Documento Fiscal</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">br_account.fiscal.document</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_type">form</field>
+        <field name="view_id" ref="view_br_account_fiscal_document_tree"/>
+    </record>
+
+    <menuitem id="menu_action_br_account_fiscal_document_form"
+              name="Documento Fiscal"
+              parent="menu_br_account"
+              action="action_br_account_fiscal_document_form"/>
+
+</odoo>
