+++ br_account/views/br_account_fiscal_observation.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record id="br_account_fiscal_observation_form" model="ir.ui.view">
+        <field name="name">br_account.fiscal.observation.form</field>
+        <field name="model">br_account.fiscal.observation</field>
+        <field name="arch" type="xml">
+            <form string="Observações Fiscais">
+                <group>
+                    <field name="name"/>
+                    <field name="document_id"/>
+                    <field name="tipo"/>
+                </group>
+                <group string="Mensagem a ser destacada no documento eletrônico">
+                    <field name="message" nolabel="1" placeholder="Utilize ${invoice.origin} para campos dinâmicos"/>
+                </group>
+            </form>
+        </field>
+    </record>
+
+    <record id="br_account_fiscal_observation_tree" model="ir.ui.view">
+        <field name="name">br_account.fiscal.observation.tree</field>
+        <field name="model">br_account.fiscal.observation</field>
+        <field name="arch" type="xml">
+            <tree string="Observações Fiscais">
+                <field name="sequence" widget="handle"/>
+                <field name="name"/>
+                <field name="document_id"/>
+            </tree>
+        </field>
+    </record>
+
+
+</odoo>
