+++ br_account/views/br_account_service_type.xml
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record id="view_br_account_service_type_form" model="ir.ui.view">
+        <field name="name">br_account.service.type.form</field>
+        <field name="model">br_account.service.type</field>
+        <field name="arch" type="xml">
+            <form string="Tipo de Serviço">
+                <sheet>
+                    <group>
+                        <field name="code"/>
+                        <field name="name"/>
+                        <field name="internal_type"/>
+                        <field name="parent_id"/>
+                        <field name="federal_nacional"/>
+                        <field name="federal_importado"/>
+                        <field name="estadual_imposto"/>
+                        <field name="municipal_imposto"/>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record id="view_br_account_service_type_tree" model="ir.ui.view">
+        <field name="name">br_account.service.type.tree</field>
+        <field name="model">br_account.service.type</field>
+        <field name="arch" type="xml">
+            <tree string="Tipo de Serviço">
+                <field name="code" select="1"/>
+                <field name="name" select="1"/>
+                <field name="parent_id" select="1"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="view_br_account_service_type_structure" model="ir.ui.view">
+        <field name="name">br_account.service.type.structure</field>
+        <field name="model">br_account.service.type</field>
+        <field name="field_parent">child_ids</field>
+        <field name="arch" type="xml">
+            <tree string="Estrutura de Tipo de Serviço">
+                <field name="code"/>
+                <field name="name"/>
+                <field name="internal_type"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="action_br_account_service_type_form" model="ir.actions.act_window">
+        <field name="name">Tipo de Serviço</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">br_account.service.type</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_type">form</field>
+        <field name="view_id" ref="view_br_account_service_type_tree"/>
+    </record>
+
+    <record id="action_br_account_service_type_structure" model="ir.actions.act_window">
+        <field name="name">Estrutura de Tipo de Serviço</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">br_account.service.type</field>
+        <field name="view_type">tree</field>
+        <field name="domain">[('parent_id','=',False)]</field>
+        <field name="view_id" ref="view_br_account_service_type_structure"/>
+    </record>
+
+    <menuitem id="menu_action_service_type_form"
+              name="Tipo de Serviço"
+              parent="menu_br_account_fiscal_manager"
+              action="action_br_account_service_type_form"
+              sequence="50"/>
+
+    <menuitem id="menu_action_service_type_structure"
+              name="Estrutura de Tipo de Serviço"
+              parent="menu_br_account_fiscal_manager"
+              action="action_br_account_service_type_structure"
+              sequence="60"/>
+
+</odoo>
