+++ br_account/views/res_company.xml
